🚀 split up app
